Consolidate FlexSN inductor code under flexsn
