--- docs/source/APIs/spikingjelly.activation_based.triton_kernel.flex_sn_inductor.rst
@@ -1,10 +1,0 @@
-spikingjelly.activation\_based.triton\_kernel.flex\_sn\_inductor package
-========================================================================
-
-HigherOrderOp & Eager Scan
----------------------------
-
-.. automodule:: spikingjelly.activation_based.triton_kernel.flex_sn_inductor.hop
-   :members:
-   :undoc-members:
-   :show-inheritance:
