+++ docs/source/APIs/spikingjelly.activation_based.triton_kernel.flexsn.inductor.rst
@@ -1,0 +1,10 @@
+spikingjelly.activation\_based.triton\_kernel.flexsn.inductor package
+======================================================================
+
+HigherOrderOp & Eager Scan
+---------------------------
+
+.. automodule:: spikingjelly.activation_based.triton_kernel.flexsn.inductor.hop
+   :members:
+   :undoc-members:
+   :show-inheritance:
