+++ spikingjelly/activation_based/triton_kernel/__init__.py
@@ -1,5 +1,5 @@
 from . import flexsn, torch2triton  # noqa
 from .compress import *
-from . import flex_sn_inductor
+from .flexsn import inductor
 from .neuron_kernel import *
 from .surrogate_kernel import *
